+++ kubejs/server_scripts/NTC/base/unify/unify_materials.js
@@ -21,6 +21,8 @@ onEvent('recipes', (event) => {
         let rod = getPreferredItemInTag(Ingredient.of(`#forge:rods/${material}`)).id;
         let plate = getPreferredItemInTag(Ingredient.of(`#forge:plates/${material}`)).id;
 
+        let wire = getPreferredItemInTag(Ingredient.of(`#forge:wires/${material}`)).id;
+
         betterend_alloys(event, material, ore, ingot);
 
         emendatus_item_melting(event, material, ore, block, ingot, nugget, gem, dust, gear, rod, plate);
@@ -54,7 +56,7 @@ onEvent('recipes', (event) => {
         thermal_metal_ore_pulverizing(event, material, ore, dust, ingot);
         thermal_gem_ore_pulverizing(event, material, ore, dust, gem, shard);
         thermal_ingot_gem_pulverizing(event, material, ingot, dust, gem);
-        thermal_metal_casting(event, material, ingot, nugget, gear, rod, plate);
+        thermal_metal_casting(event, material, ingot, nugget, gear, rod, plate, wire);
         thermal_metal_melting(event, material, block, ingot, nugget, gear, rod, plate);
         thermal_gem_casting(event, material, gem, gear, rod, plate);
         thermal_gem_melting(event, material, block, gem, gear, rod, plate);
@@ -667,8 +669,8 @@ onEvent('recipes', (event) => {
         event.recipes.thermal.pulverizer(output, input);
     }
 
-    function thermal_metal_casting(event, material, ingot, nugget, gear, rod, plate) {
-        if (material == air || ingot == air || nugget == air || gear == air || rod == air || plate == air) {
+    function thermal_metal_casting(event, material, ingot, nugget, gear, rod, plate, wire) {
+        if (material == air || ingot == air || nugget == air || gear == air || rod == air || plate == air || wire == air) {
             return;
         }
 
@@ -697,6 +699,9 @@ onEvent('recipes', (event) => {
         }
         if (plate != air) {
             recipes.push({ type: 'plate', amount: 144, output: plate, energy: 5000 });
+        }
+        if (wire != air) {
+            recipes.push({ type: 'wire', amount: 72, output: wire, energy: 2500 });
         }
 
         recipes.forEach((recipe) => {